requirements.txt
langchain
ipykernel
python-dotenv
langchain-community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-openai
sentence-transformers
huggingface_hub
langchain-huggingface
faiss-cpu
chromadb
langchain_chroma
beautifulsoup4
streamlit
langchain_groq
langserve
fastapi
sse_starlette
bs4
